move zome and links
