+++ Cargo.toml
@@ -23,6 +23,7 @@ members = [
   "crates/holochain_websocket",
   "crates/holochain_util",
   "crates/holochain_zome_types",
+  "crates/holochain_integrity_types",
 
   "crates/kitsune_p2p/dht_arc",
   "crates/kitsune_p2p/bootstrap",
