+++ crates/holochain_integrity_types/Cargo.toml
@@ -1,0 +1,43 @@
+[package]
+name = "holochain_integrity_types"
+version = "0.0.1"
+description = "Holochain integrity types"
+license = "Apache-2.0"
+homepage = "https://github.com/holochain/holochain"
+documentation = "https://docs.rs/holochain_integrity_types"
+readme = "README.md"
+authors = ["Holochain Core Dev Team <devcore@holochain.org>"]
+edition = "2021"
+
+[dependencies]
+holo_hash = { version = "0.0.21", path = "../holo_hash" }
+holochain_serialized_bytes = "=0.0.51"
+serde = { version = "1.0", features = ["derive"] }
+serde_bytes = "0.11"
+
+# Just the bare minimum timestamp with no extra features.
+# TODO: This needs to point to a published version of this crate and be pinned.
+kitsune_p2p_timestamp = { version = "0.0.7", path = "../kitsune_p2p/timestamp", default-features = false }
+
+# TODO: Figure out how to remove these dependencies.
+subtle = "2"
+
+# Optional
+tracing = { version = "0.1", optional = true }
+arbitrary = { version = "1.0", features = ["derive"], optional = true }
+
+[features]
+default = []
+
+full = ["default", "kitsune_p2p_timestamp/full"]
+
+test_utils = [
+  "arbitrary",
+  "kitsune_p2p_timestamp/arbitrary",
+  "kitsune_p2p_timestamp/now",
+  "kitsune_p2p_timestamp/rusqlite",
+  "holo_hash/arbitrary",
+  "holo_hash/hashing",
+  "holo_hash/test_utils",
+  "holochain_serialized_bytes/arbitrary",
+]
