+++ crates/holochain_zome_types/Cargo.toml
@@ -12,8 +12,8 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-chrono = "0.4.6"
 kitsune_p2p_timestamp = { version = "0.0.7", path = "../kitsune_p2p/timestamp" }
+holochain_integrity_types = { version = "0.0.1", path = "../holochain_integrity_types", features = ["tracing"] }
 holo_hash = { version = "0.0.21", path = "../holo_hash" }
 holochain_serialized_bytes = "=0.0.51"
 paste = "=1.0.5"
@@ -77,5 +77,6 @@ test_utils = [
   "holo_hash/hashing",
   "holo_hash/test_utils",
   "holochain_serialized_bytes/arbitrary",
-  "full-dna-def"
+  "full-dna-def",
+  "holochain_integrity_types/test_utils",
 ]
